refactor: alleviate circular import additions
