+++ packages/insomnia-inso/src/db/adapters/insomnia-adapter.ts
@@ -5,9 +5,9 @@ import { importInsomniaV5Data } from 'insomnia/src/common/insomnia-v5';
 import YAML from 'yaml';
 
 import { InsoError } from '../../errors';
-import type { DbAdapter } from '../index';
-import { emptyDb } from '../index';
 import type { BaseModel } from '../models/types';
+import type { DbAdapter } from '../types';
+import { emptyDb } from '../types';
 
 /**
  * When exporting from Insomnia, the `models.[kind].type` is converted from PascalCase to snake_case.
