+++ packages/insomnia-inso/src/db/index.ts
@@ -1,58 +1,14 @@
 import { stat } from 'node:fs/promises';
 
-import type { CaCertificate } from 'insomnia/src/models/ca-certificate';
-import type { ClientCertificate } from 'insomnia/src/models/client-certificate';
-import type { CloudProviderCredential } from 'insomnia/src/models/cloud-credential';
-import type { CookieJar } from 'insomnia/src/models/cookie-jar';
-import type { Settings } from 'insomnia/src/models/settings';
-
 import { logger } from '../logger';
 import gitAdapter from './adapters/git-adapter';
 import insomniaExportAdapter from './adapters/insomnia-adapter';
 import neDbAdapter from './adapters/ne-db-adapter';
-import type {
-  ApiSpec,
-  BaseModel,
-  Environment,
-  UnitTest,
-  UnitTestSuite,
-  Workspace,
-  WorkspaceMeta,
-} from './models/types';
+import type { Database } from './types';
+import { emptyDb } from './types';
 
-export interface Database {
-  ApiSpec: ApiSpec[];
-  Environment: Environment[];
-  Request: BaseModel[];
-  RequestGroup: BaseModel[];
-  Workspace: Workspace[];
-  WorkspaceMeta: WorkspaceMeta[];
-  UnitTestSuite: UnitTestSuite[];
-  UnitTest: UnitTest[];
-  ClientCertificate: ClientCertificate[];
-  CaCertificate: CaCertificate[];
-  CookieJar: CookieJar[];
-  CloudCredential: CloudProviderCredential[];
-  Settings: Settings[];
-}
-
-export const emptyDb = (): Database => ({
-  ApiSpec: [],
-  Environment: [],
-  Request: [],
-  RequestGroup: [],
-  Workspace: [],
-  WorkspaceMeta: [],
-  UnitTest: [],
-  UnitTestSuite: [],
-  ClientCertificate: [],
-  CaCertificate: [],
-  CookieJar: [],
-  CloudCredential: [],
-  Settings: [],
-});
-
-export type DbAdapter = (dir: string, filterTypes?: (keyof Database)[]) => Promise<Database | null>;
+export type { Database, DbAdapter } from './types';
+export { emptyDb };
 
 interface Options {
   pathToSearch: string;
